Move StructureFilterSet to the right app
