+++ geotrek/authent/filters.py
@@ -1,0 +1,6 @@
+from mapentity.filters import MapEntityFilterSet
+
+
+class StructureRelatedFilterSet(MapEntityFilterSet):
+    class Meta(MapEntityFilterSet.Meta):
+        fields = MapEntityFilterSet.Meta.fields + ['structure']
